+++ api/src/pdc/services/policy/engine/policies/index.ts
@@ -16,7 +16,7 @@ import { NantesMetropole2024 } from './20240101_NantesMetropole';
 import { Nm } from './20211202_NantesMetropoleXP';
 import { Occitanie20232024 } from './20221024_Occitanie';
 import { PMGFxATMB2024 } from './PMGFxATMB2024';
-import { PaysBasque20232024 } from './PaysBasque20232024';
+import { PaysBasque20232024 } from './20230401_PaysBasqueAdour';
 import { PaysDeLaLoire2021 } from './PaysDeLaLoire2021';
 import { PaysDeLaLoire2023 } from './PaysDeLaLoire2023';
 import { PaysDeLaLoire2024 } from './PaysDeLaLoire2024';
